requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.3
PyYAML==6.0.1
python-box==7.0.1
ensure==1.0.2
dill==0.3.7
imbalanced-learn==0.11.0
xgboost==1.7.6
evidently==0.4.5
pytest==7.4.0
